requirements.txt
redis
djangorestframework==3.11.0
Django==3.0.7
gunicorn==20.0.4
django-cors-headers